refactor: clean and re-structure
