--- src/render/core/app/components/app-movie-details-info/index.css
@@ -1,3 +1,0 @@
-
-
-/*# sourceMappingURL=index.css.map */
