+++ src/render/core/app/components/app-movie-details/index.jsx
