--- src/render/core/app/layout/app-movie-details/index.css
@@ -1,3 +1,0 @@
-
-
-/*# sourceMappingURL=index.css.map */
