--- src/render/core/app/layout/app-movies-list/index.css
@@ -1,3 +1,0 @@
-
-
-/*# sourceMappingURL=index.css.map */
