+++ src/render/core/app/pages/index-view/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import AppMovies from 'layout/app-movies-list/'
-import AppMovieDetails from 'layout/app-movie-details/'
+import AppMovies from 'components/app-movies-list/'
+import AppMovieDetails from 'components/app-movie-details/'
 import AppNav from 'components/app-movies-nav-bar/'
 import AppSearch from 'components/app-movies-search/'
 import AppLoader from 'components/app-movie-player-loader/'
